backend/requirements.txt
openai 
python-dotenv
sounddevice
scipy
keyboard
textblob
fastapi
uvicorn[standard]
websockets
numpy